74-grpc-service/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
protobuf
pydantic
